+++ InvenTree/InvenTree/tasks.py
@@ -12,7 +12,6 @@ from datetime import datetime, timedelta
 from typing import Callable, List
 
 from django.conf import settings
-from django.core import mail as django_mail
 from django.core.exceptions import AppRegistryNotReady
 from django.core.management import call_command
 from django.db import DEFAULT_DB_ALIAS, connections
@@ -557,28 +556,6 @@ def run_backup():
 
     # Record that this task was successful
     record_task_success('run_backup')
-
-
-def send_email(subject, body, recipients, from_email=None, html_message=None):
-    """Send an email with the specified subject and body, to the specified recipients list."""
-    if type(recipients) == str:
-        recipients = [recipients]
-
-    import InvenTree.ready
-
-    if InvenTree.ready.isImportingData():
-        # If we are importing data, don't send emails
-        return
-
-    offload_task(
-        django_mail.send_mail,
-        subject,
-        body,
-        from_email,
-        recipients,
-        fail_silently=False,
-        html_message=html_message
-    )
 
 
 @scheduled_task(ScheduledTask.DAILY)
